Add Int64Index, Float64Index, DatetimeIndex.
